+++ platform/ios/src/MGLCalloutView.h
@@ -1,0 +1,6 @@
+#import "SMCalloutView.h"
+#import "MGLCalloutViewProtocol.h"
+
+@interface MGLCalloutView : SMCalloutView <MGLCalloutViewProtocol>
+
+@end